+++ CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.12.5] - 2024-08-27
+## [0.12.5] - 2024-08-29
 
 ### Added
 - Added support for creating a client using an OIDC token (CNJR-6081)
 - Added support for rotating the API key of an authenticated host (CNJR-5699)
+- Added support for checking authenticator status (CNJR-6371)
 - Return error when trying to call unsupported methods in Conjur Cloud (CNJR-5700)
 
 ## [0.12.4] - 2024-08-01